Feat : 일정 수정 및 삭제시 비밀번호 검증
일정 수정 및 삭제 시도할때 비밀번호가 다르면 403 FORBIDDEN 이 뜨도록 추가

+++ src/main/java/com/example/schedulemanagement/service/ScheduleServiceImpl.java
@@ -46,27 +46,35 @@ public class ScheduleServiceImpl implements ScheduleService {
     }
 
     @Override
-    public ScheduleResponseDto updateSchedule(Long id, String userName, String title, String contents) {
+    public ScheduleResponseDto updateSchedule(Long id, ScheduleRequestDto dto) {
         Schedule schedule = scheduleRepository.findScheduleById(id);
 
         if (schedule == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Schedule not found");
         }
-        if (userName == null || title == null) {
+        if (dto.getUserName() == null || dto.getTitle() == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Username and title cannot be null");
         }
+        if (!schedule.getPassword().equals(dto.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Passwords don't match");
+        }
 
-        schedule.update(userName, title, contents);
+
+        schedule.update(dto.getUserName(), dto.getTitle(), dto.getContents());
 
         return new ScheduleResponseDto(schedule);
     }
 
     @Override
-    public void deleteSchedule(Long id) {
+    public void deleteSchedule(Long id, ScheduleRequestDto dto) {
         Schedule schedule = scheduleRepository.findScheduleById(id);
 
         if (schedule == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Schedule not found");
+        }
+
+        if (!schedule.getPassword().equals(dto.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Passwords don't match");
         }
 
         scheduleRepository.deleteSchedule(id);
